feat: removed client button from home
