+++ src/app/style/lib/_themes.scss
@@ -4,53 +4,53 @@ $themes: (
     // Light theme
     light: (
         // Primary and accent colors
-        primary:    #496F5D,
-        primary2:   #63907B,
-        primary3:   #6EAA8D,
-        primary4:   #74C69F,
-        primary5:   #9AF6CB,
-        accent:     #52796F,
-        accent2:    #275549,
-        accent3:    #11382E,
+        p:    #496F5D,
+        p2:   #63907B,
+        p3:   #6EAA8D,
+        p4:   #74C69F,
+        p5:   #9AF6CB,
+        a:     #52796F,
+        a2:    #275549,
+        a3:    #11382E,
         
         // Secondary colors
-        secondary:  #3F334D,
-        secondary2: #635276,
-        secondary3: #8D74A9,
-        secondary4: #8D74A9,
+        s:  #3F334D,
+        s2: #635276,
+        s3: #8D74A9,
+        s4: #8D74A9,
 
         // Neutral colors
-        neutral:    #EEF0F2,
-        neutral2:   #C6C7C4,
-        neutral3:   #9C9C9C,
-        neutral4:   #3D3D3D,
-        neutral5:   #0E0E0E,
+        n:    #EEF0F2,
+        n2:   #C6C7C4,
+        n3:   #9C9C9C,
+        n4:   #3D3D3D,
+        n5:   #0E0E0E,
     ),
 
     // Dark theme
     dark: (
         // Primary and accent colors
-        primary:    #000,
-        primary2:   #000,
-        primary3:   #000,
-        primary4:   #000,
-        primary5:   #000,
-        accent:     #000,
-        accent2:    #000,
-        accent3:    #000,
+        p:    #000,
+        p2:   #000,
+        p3:   #000,
+        p4:   #000,
+        p5:   #000,
+        a:     #000,
+        a2:    #000,
+        a3:    #000,
         
         // Secondary colors
-        secondary:  #000,
-        secondary2: #000,
-        secondary3: #000,
-        secondary4: #000,
+        s:  #000,
+        s2: #000,
+        s3: #000,
+        s4: #000,
 
         // Neutral colors
-        neutral:    #000,
-        neutral2:   #000,
-        neutral3:   #000,
-        neutral4:   #000,
-        neutral5:   #000,
+        n:    #000,
+        n2:   #000,
+        n3:   #000,
+        n4:   #000,
+        n5:   #000,
     )
 );
