+++ src/app/style/pages/_index.scss
@@ -2,6 +2,6 @@
 
 div {
     @include themed() {
-        background: c(primary);
+        background: c(p);
     }
 }
